split `method_bind_common.h` from `method_bind.h`
