+++ core/core_constants.cpp
@@ -31,7 +31,7 @@
 #include "core_constants.h"
 
 #include "core/input/input_enums.h"
-#include "core/object/method_bind.h" // IWYU pragma: keep. To bind `MethodFlags`.
+#include "core/object/method_bind_common.h" // IWYU pragma: keep. For `VARIANT_BITFIELD_CAST(MethodFlags)`.
 #include "core/object/object.h"
 #include "core/os/keyboard.h"
 #include "core/variant/type_info.h"
