+++ core/object/method_bind.cpp
@@ -28,10 +28,9 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
+#include "method_bind.h"
+
 #include "core/object/object.h"
-// object.h needs to be the first include *before* method_bind.h
-// FIXME: Find out why and fix potential cyclical dependencies.
-#include "core/object/method_bind.h"
 
 uint32_t MethodBind::get_hash() const {
 	MethodInfo mi;
